requirements.txt
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.4
anyio==4.4.0
astor==0.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
contourpy==1.2.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
distlib==0.3.8
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.7
imageio==2.35.0
imgaug==0.4.0
jmespath==0.10.0
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
matplotlib==3.9.2
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
oss2==2.18.6
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
protobuf==5.27.3
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
requests==2.32.3
retrying==1.3.4
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
termcolor==2.4.0
tifffile==2024.8.10
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
img2table==1.2.11